Reorganise dopri5 code
